+++ e2e/contracts/sui/bin.go
@@ -14,6 +14,12 @@ var fakeUSDC []byte
 //go:embed evm.mv
 var evmBinary []byte
 
+//go:embed token.mv
+var tokenBinary []byte
+
+//go:embed connected.mv
+var connectedBinary []byte
+
 // GatewayBytecodeBase64 gets the gateway binary encoded as base64 for deployment
 func GatewayBytecodeBase64() string {
 	return base64.StdEncoding.EncodeToString(gatewayBinary)
@@ -27,4 +33,14 @@ func FakeUSDCBytecodeBase64() string {
 // EVMBytecodeBase64 gets the EVM binary encoded as base64 for deployment
 func EVMBytecodeBase64() string {
 	return base64.StdEncoding.EncodeToString(evmBinary)
+}
+
+// ExampleFungibleTokenBytecodeBase64 gets the example package's fungible token binary encoded as base64 for deployment
+func ExampleFungibleTokenBytecodeBase64() string {
+	return base64.StdEncoding.EncodeToString(tokenBinary)
+}
+
+// ExampleConnectedBytecodeBase64 gets the example package's connected binary encoded as base64 for deployment
+func ExampleConnectedBytecodeBase64() string {
+	return base64.StdEncoding.EncodeToString(connectedBinary)
 }