+++ zetaclient/chains/sui/client/client.go
@@ -10,7 +10,10 @@ import (
 
 	"github.com/block-vision/sui-go-sdk/models"
 	"github.com/block-vision/sui-go-sdk/sui"
+	suiptb "github.com/pattonkan/sui-go/sui"
 	"github.com/pkg/errors"
+
+	zetasui "github.com/zeta-chain/node/pkg/contracts/sui"
 )
 
 // Client Sui client.
@@ -247,6 +250,60 @@ func (c *Client) SuiExecuteTransactionBlock(
 	return parseRPCResponse[models.SuiTransactionBlockResponse]([]byte(resString))
 }
 
+// GetSuiCoinObjectRef returns the latest SUI coin object reference for given owner address
+// Note: the SUI object may change over time, so we need to get the latest object
+func (c *Client) GetSuiCoinObjectRef(ctx context.Context, owner string) (suiptb.ObjectRef, error) {
+	coins, err := c.SuiXGetCoins(ctx, models.SuiXGetCoinsRequest{
+		Owner:    owner,
+		CoinType: string(zetasui.SUI),
+	})
+	if err != nil {
+		return suiptb.ObjectRef{}, errors.Wrap(err, "unable to get TSS coins")
+	}
+
+	var (
+		suiCoin        *models.CoinData
+		suiCoinVersion uint64
+	)
+
+	// locate the latest version of SUI coin object of given owner
+	for _, coin := range coins.Data {
+		if !zetasui.IsSUICoinType(zetasui.CoinType(coin.CoinType)) {
+			continue
+		}
+
+		version, err := strconv.ParseUint(coin.Version, 10, 64)
+		if err != nil {
+			return suiptb.ObjectRef{}, errors.Wrapf(err, "failed to parse SUI coin version %s", coin.Version)
+		}
+
+		if version > suiCoinVersion {
+			suiCoin = &coin
+			suiCoinVersion = version
+		}
+	}
+	if suiCoin == nil {
+		return suiptb.ObjectRef{}, errors.New("SUI coin not found")
+	}
+
+	// convert coin data to object ref
+	suiCoinID, err := suiptb.ObjectIdFromHex(suiCoin.CoinObjectId)
+	if err != nil {
+		return suiptb.ObjectRef{}, errors.Wrapf(err, "failed to parse SUI coin ID: %s", suiCoin.CoinObjectId)
+	}
+
+	suiCoinDigest, err := suiptb.NewBase58(suiCoin.Digest)
+	if err != nil {
+		return suiptb.ObjectRef{}, errors.Wrapf(err, "failed to parse SUI coin digest: %s", suiCoin.Digest)
+	}
+
+	return suiptb.ObjectRef{
+		ObjectId: suiCoinID,
+		Version:  suiCoinVersion,
+		Digest:   suiCoinDigest,
+	}, nil
+}
+
 // EncodeCursor encodes event ID into cursor.
 func EncodeCursor(id models.EventId) string {
 	return fmt.Sprintf("%s,%s", id.TxDigest, id.EventSeq)
@@ -317,28 +374,29 @@ func (c *Client) CheckObjectIDsShared(ctx context.Context, objectIDs []string) e
 		return fmt.Errorf("expected %d objects, but got %d", len(objectIDs), len(res))
 	}
 
-	return checkContainOwnedObject(res)
+	return CheckContainOwnedObject(res)
 }
 
-func checkContainOwnedObject(res []*models.SuiObjectResponse) error {
+// CheckContainOwnedObject checks if the provided object list represents Sui shared or immmutable objects
+func CheckContainOwnedObject(res []*models.SuiObjectResponse) error {
 	for i, obj := range res {
 		if obj.Data == nil {
-			return fmt.Errorf("object %d is missing data", i)
+			return errors.Wrapf(zetasui.ErrObjectOwnership, "object %d is missing data", i)
 		}
 
 		switch owner := obj.Data.Owner.(type) {
 		case string:
 			if owner != immutableOwner {
-				return fmt.Errorf("object %d has unexpected string owner: %s", i, owner)
+				return errors.Wrapf(zetasui.ErrObjectOwnership, "object %d has unexpected string owner: %s", i, owner)
 			}
 			// Immutable is valid, continue
-		case map[string]interface{}:
+		case map[string]any:
 			if _, isShared := owner[sharedOwner]; !isShared {
-				return fmt.Errorf("object %d is not shared or immutable: owner = %+v", i, owner)
+				return errors.Wrapf(zetasui.ErrObjectOwnership, "object %d is not shared or immutable: owner = %+v", i, owner)
 			}
 			// Shared is valid, continue
 		default:
-			return fmt.Errorf("object %d has unknown owner type: %+v", i, obj.Data.Owner)
+			return errors.Wrapf(zetasui.ErrObjectOwnership, "object %d has unknown owner type: %+v", i, obj.Data.Owner)
 		}
 	}