+++ changelog.md
@@ -22,6 +22,7 @@
 * [3797](https://github.com/zeta-chain/node/pull/3797) - integrate execute SPL revert
 * [3807](https://github.com/zeta-chain/node/pull/3807) - integrate ZEVM to Solana call
 * [3826](https://github.com/zeta-chain/node/pull/3826) - add global tss signature rate-limiter to zetaclient
+* [3793](https://github.com/zeta-chain/node/pull/3793) - support Sui withdrawAndCall using the PTB transaction
 
 ### Refactor